requirements.txt
streamlit
requests
textblob
python-dotenv